fix(db): 掃除の件数を、要らない変換なしで足す
`answer.expiring` は既に `u64` なので `u64::from` は無用だった。clippy が
`useless conversion` として落とす。

**軽い検査（`cargo check`）では通っていた。** 型としては正しいので、
**`make ci` の clippy まで来ないと出ない**——「部分実行の緑」の一例である。

+++ server/crates/server-core/src/db/memo_blobs.rs
@@ -176,7 +176,7 @@ async fn sweep_expired(
         // **`sweep_bytes` は 0。** 容量のぶんは同意を取ってからでないと消せない
         let answer =
             survey_or_sweep(db, account_id, now_ms, days, limits.max_bytes, 0, true).await?;
-        removed += u64::from(answer.expiring);
+        removed += answer.expiring;
     }
     Ok(removed)
 }
